Allow for custom exception handlers.
