+++ system/Debug/Exceptions.php
@@ -11,12 +11,9 @@
 
 namespace CodeIgniter\Debug;
 
-use CodeIgniter\API\ResponseTrait;
-use CodeIgniter\Exceptions\PageNotFoundException;
 use CodeIgniter\HTTP\IncomingRequest;
 use CodeIgniter\HTTP\Response;
 use Config\Exceptions as ExceptionsConfig;
-use Config\Paths;
 use ErrorException;
 use Throwable;
 
@@ -25,23 +22,6 @@ use Throwable;
  */
 class Exceptions
 {
-    use ResponseTrait;
-
-    /**
-     * Nesting level of the output buffering mechanism
-     *
-     * @var int
-     */
-    public $ob_level;
-
-    /**
-     * The path to the directory containing the
-     * cli and html error view directories.
-     *
-     * @var string
-     */
-    protected $viewPath;
-
     /**
      * Config for debug exceptions.
      *
@@ -65,8 +45,6 @@ class Exceptions
 
     public function __construct(ExceptionsConfig $config, IncomingRequest $request, Response $response)
     {
-        $this->ob_level = ob_get_level();
-        $this->viewPath = rtrim($config->errorViewPath, '\\/ ') . DIRECTORY_SEPARATOR;
         $this->config   = $config;
         $this->request  = $request;
         $this->response = $response;
@@ -107,20 +85,25 @@ class Exceptions
             ]);
         }
 
-        if (! is_cli()) {
-            $this->response->setStatusCode($statusCode);
-            header(sprintf('HTTP/%s %s %s', $this->request->getProtocolVersion(), $this->response->getStatusCode(), $this->response->getReasonPhrase()), true, $statusCode);
+        $config = config('Exceptions');
 
-            if (strpos($this->request->getHeaderLine('accept'), 'text/html') === false) {
-                $this->respond(ENVIRONMENT === 'development' ? $this->collectVars($exception, $statusCode) : '', $statusCode)->send();
-
-                exit($exitCode);
-            }
+        if (! method_exists($config, 'handler')) {
+            exit('Config\Exception must have a handler() method.');
         }
 
-        $this->render($exception, $statusCode);
+        $handler = config('Exceptions')->handler($statusCode, $exception);
 
-        exit($exitCode);
+        if (! $handler instanceof BaseExceptionHandler) {
+            exit('Exception Handler not found.');
+        }
+
+        $handler
+            ->setException($exception)
+            ->setStatusCode($statusCode)
+            ->setExitCode($exitCode)
+            ->setRequest($this->request)
+            ->setResponse($this->response)
+            ->handle();
     }
 
     /**
@@ -165,132 +148,6 @@ class Exceptions
     }
 
     /**
-     * Determines the view to display based on the exception thrown,
-     * whether an HTTP or CLI request, etc.
-     *
-     * @return string The path and filename of the view file to use
-     */
-    protected function determineView(Throwable $exception, string $templatePath): string
-    {
-        // Production environments should have a custom exception file.
-        $view         = 'production.php';
-        $templatePath = rtrim($templatePath, '\\/ ') . DIRECTORY_SEPARATOR;
-
-        if (str_ireplace(['off', 'none', 'no', 'false', 'null'], '', ini_get('display_errors'))) {
-            $view = 'error_exception.php';
-        }
-
-        // 404 Errors
-        if ($exception instanceof PageNotFoundException) {
-            return 'error_404.php';
-        }
-
-        // Allow for custom views based upon the status code
-        if (is_file($templatePath . 'error_' . $exception->getCode() . '.php')) {
-            return 'error_' . $exception->getCode() . '.php';
-        }
-
-        return $view;
-    }
-
-    /**
-     * Given an exception and status code will display the error to the client.
-     */
-    protected function render(Throwable $exception, int $statusCode)
-    {
-        // Determine possible directories of error views
-        $path    = $this->viewPath;
-        $altPath = rtrim((new Paths())->viewDirectory, '\\/ ') . DIRECTORY_SEPARATOR . 'errors' . DIRECTORY_SEPARATOR;
-
-        $path    .= (is_cli() ? 'cli' : 'html') . DIRECTORY_SEPARATOR;
-        $altPath .= (is_cli() ? 'cli' : 'html') . DIRECTORY_SEPARATOR;
-
-        // Determine the views
-        $view    = $this->determineView($exception, $path);
-        $altView = $this->determineView($exception, $altPath);
-
-        // Check if the view exists
-        if (is_file($path . $view)) {
-            $viewFile = $path . $view;
-        } elseif (is_file($altPath . $altView)) {
-            $viewFile = $altPath . $altView;
-        }
-
-        if (! isset($viewFile)) {
-            echo 'The error view files were not found. Cannot render exception trace.';
-
-            exit(1);
-        }
-
-        if (ob_get_level() > $this->ob_level + 1) {
-            ob_end_clean();
-        }
-
-        echo(function () use ($exception, $statusCode, $viewFile): string {
-            $vars = $this->collectVars($exception, $statusCode);
-            extract($vars, EXTR_SKIP);
-
-            ob_start();
-            include $viewFile;
-
-            return ob_get_clean();
-        })();
-    }
-
-    /**
-     * Gathers the variables that will be made available to the view.
-     */
-    protected function collectVars(Throwable $exception, int $statusCode): array
-    {
-        $trace = $exception->getTrace();
-
-        if ($this->config->sensitiveDataInTrace !== []) {
-            $this->maskSensitiveData($trace, $this->config->sensitiveDataInTrace);
-        }
-
-        return [
-            'title'   => get_class($exception),
-            'type'    => get_class($exception),
-            'code'    => $statusCode,
-            'message' => $exception->getMessage() ?? '(null)',
-            'file'    => $exception->getFile(),
-            'line'    => $exception->getLine(),
-            'trace'   => $trace,
-        ];
-    }
-
-    /**
-     * Mask sensitive data in the trace.
-     *
-     * @param array|object $trace
-     */
-    protected function maskSensitiveData(&$trace, array $keysToMask, string $path = '')
-    {
-        foreach ($keysToMask as $keyToMask) {
-            $explode = explode('/', $keyToMask);
-            $index   = end($explode);
-
-            if (strpos(strrev($path . '/' . $index), strrev($keyToMask)) === 0) {
-                if (is_array($trace) && array_key_exists($index, $trace)) {
-                    $trace[$index] = '******************';
-                } elseif (is_object($trace) && property_exists($trace, $index) && isset($trace->{$index})) {
-                    $trace->{$index} = '******************';
-                }
-            }
-        }
-
-        if (is_object($trace)) {
-            $trace = get_object_vars($trace);
-        }
-
-        if (is_array($trace)) {
-            foreach ($trace as $pathKey => $subarray) {
-                $this->maskSensitiveData($subarray, $keysToMask, $path . '/' . $pathKey);
-            }
-        }
-    }
-
-    /**
      * Determines the HTTP status code and the exit status code for this request.
      */
     protected function determineCodes(Throwable $exception): array
@@ -310,124 +167,5 @@ class Exceptions
         }
 
         return [$statusCode, $exitStatus];
-    }
-
-    //--------------------------------------------------------------------
-    // Display Methods
-    //--------------------------------------------------------------------
-
-    /**
-     * This makes nicer looking paths for the error output.
-     */
-    public static function cleanPath(string $file): string
-    {
-        switch (true) {
-            case strpos($file, APPPATH) === 0:
-                $file = 'APPPATH' . DIRECTORY_SEPARATOR . substr($file, strlen(APPPATH));
-                break;
-
-            case strpos($file, SYSTEMPATH) === 0:
-                $file = 'SYSTEMPATH' . DIRECTORY_SEPARATOR . substr($file, strlen(SYSTEMPATH));
-                break;
-
-            case strpos($file, FCPATH) === 0:
-                $file = 'FCPATH' . DIRECTORY_SEPARATOR . substr($file, strlen(FCPATH));
-                break;
-
-            case defined('VENDORPATH') && strpos($file, VENDORPATH) === 0:
-                $file = 'VENDORPATH' . DIRECTORY_SEPARATOR . substr($file, strlen(VENDORPATH));
-                break;
-        }
-
-        return $file;
-    }
-
-    /**
-     * Describes memory usage in real-world units. Intended for use
-     * with memory_get_usage, etc.
-     */
-    public static function describeMemory(int $bytes): string
-    {
-        if ($bytes < 1024) {
-            return $bytes . 'B';
-        }
-
-        if ($bytes < 1_048_576) {
-            return round($bytes / 1024, 2) . 'KB';
-        }
-
-        return round($bytes / 1_048_576, 2) . 'MB';
-    }
-
-    /**
-     * Creates a syntax-highlighted version of a PHP file.
-     *
-     * @return bool|string
-     */
-    public static function highlightFile(string $file, int $lineNumber, int $lines = 15)
-    {
-        if (empty($file) || ! is_readable($file)) {
-            return false;
-        }
-
-        // Set our highlight colors:
-        if (function_exists('ini_set')) {
-            ini_set('highlight.comment', '#767a7e; font-style: italic');
-            ini_set('highlight.default', '#c7c7c7');
-            ini_set('highlight.html', '#06B');
-            ini_set('highlight.keyword', '#f1ce61;');
-            ini_set('highlight.string', '#869d6a');
-        }
-
-        try {
-            $source = file_get_contents($file);
-        } catch (Throwable $e) {
-            return false;
-        }
-
-        $source = str_replace(["\r\n", "\r"], "\n", $source);
-        $source = explode("\n", highlight_string($source, true));
-        $source = str_replace('<br />', "\n", $source[1]);
-        $source = explode("\n", str_replace("\r\n", "\n", $source));
-
-        // Get just the part to show
-        $start = max($lineNumber - (int) round($lines / 2), 0);
-
-        // Get just the lines we need to display, while keeping line numbers...
-        $source = array_splice($source, $start, $lines, true);
-
-        // Used to format the line number in the source
-        $format = '% ' . strlen((string) ($start + $lines)) . 'd';
-
-        $out = '';
-        // Because the highlighting may have an uneven number
-        // of open and close span tags on one line, we need
-        // to ensure we can close them all to get the lines
-        // showing correctly.
-        $spans = 1;
-
-        foreach ($source as $n => $row) {
-            $spans += substr_count($row, '<span') - substr_count($row, '</span');
-            $row = str_replace(["\r", "\n"], ['', ''], $row);
-
-            if (($n + $start + 1) === $lineNumber) {
-                preg_match_all('#<[^>]+>#', $row, $tags);
-
-                $out .= sprintf(
-                    "<span class='line highlight'><span class='number'>{$format}</span> %s\n</span>%s",
-                    $n + $start + 1,
-                    strip_tags($row),
-                    implode('', $tags[0])
-                );
-            } else {
-                $out .= sprintf('<span class="line"><span class="number">' . $format . '</span> %s', $n + $start + 1, $row) . "\n";
-            }
-        }
-
-        if ($spans > 0) {
-            $out .= str_repeat('</span>', $spans);
-        }
-
-        return '<pre><code>' . $out . '</code></pre>';
     }
 }
